Moving image_loader func to loader.py
